Extracts `HackrNewsFeedStoreSpecs` helper methods into a shared scope
